Add a field Overtime payment time to the timesheet view
It should be shown below the Total time field.

Total time should show the sum of all real hours that the employee has worked (and registered) for this timesheet period.

The field Overtime payment time shows the sum of all hours written on the project Overtime payment.
Normally, we expect hours on this project to be written not more than one time.
Furthermore we expect a negative amount (allthough this is not checked).

The field is not shown if there aren't any hours written on the project Overtime payment.