fix flexitime balance miscalculations
total_duty_hours_done is not being computed, total_duty_hours is. When the timesheet is still open that is the field we want to use, so the actual value is always computed. Important changes that require recomputing: time off requests of user U AFTER creation of timesheet T for same user U